Starting working on notification sending logic.
