+++ coin-alert/src/app/lib/utils/solanaUtils.ts
@@ -1,0 +1,14 @@
+import { TokenAmount } from "@solana/web3.js";
+
+export interface TokenAccountData {
+    info: TokenAccountInfo
+    "type": string;
+}
+
+export interface TokenAccountInfo {
+    isNative: boolean;
+    mint: string;
+    owner: string;
+    state: string;
+    tokenAmount: TokenAmount;
+}
